+++ collector-amazon-ivs/src/androidTest/java/com/bitmovin/analytics/amazon/ivs/IvsPlayerConstants.kt
@@ -1,0 +1,5 @@
+package com.bitmovin.analytics.amazon.ivs
+
+object IvsPlayerConstants {
+    val playerInfo = PlayerInfo("amazonivs", "amazonivs-1.17.0", "Android:AmazonIVS")
+}