+++ collector-amazon-ivs/src/androidTest/java/com/bitmovin/analytics/amazon/ivs/PlayerInfo.kt
@@ -1,0 +1,3 @@
+package com.bitmovin.analytics.amazon.ivs
+
+data class PlayerInfo(val playerName: String, val playerVersion: String, val playerTech: String)