+++ collector-amazon-ivs/src/androidTest/java/com/bitmovin/analytics/amazon/ivs/PlayerSettings.kt
@@ -1,0 +1,3 @@
+package com.bitmovin.analytics.amazon.ivs
+
+data class PlayerSettings(val isMuted: Boolean)