+++ collector-amazon-ivs/src/androidTest/java/com/bitmovin/analytics/amazon/ivs/StreamData.kt
@@ -1,0 +1,3 @@
+package com.bitmovin.analytics.amazon.ivs
+
+data class StreamData(val videoCodec: String, val audioCodec: String, val m3u8Url: String, val streamFormat: String, val isLive: Boolean, val duration: Long)